Basic view support/Bug fix
On the month, basicWeek and basicDay views right clicking on a day didn't always raise the right click event and on the agendaWeek and agendaDay views right clicking on the all day events wasn't supported. To support this I modified the monkeyPatchViewClass to support multiple parameters and build the css selector to bind to. After navigating to a new month, week or day multiple right click events get raised. To correct this I call 'off' on the contextmenu event before binding it.